make it easy for Hibernate Reactive to reuse the @NaturalId stuff
